Automatic
In this version of the code I made it so that the code ran itself after every song change, using time remaining. I had to remove the pause/play part of the code because it didn't work with the rest of the code (to my knowledge), and made another file with just that part of the code.